+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix eslintignore not ignorning package files ([#1486](https://github.com/react-static/react-static/pull/1486))
 - Remove `@types/react-hot-loader` from TypeScript template ([#1485](https://github.com/react-static/react-static/pull/1485))
 - Expand `styled-components` peer dependency version range in `react-static-plugin-styled-components` to allow newer versions of styled-components to be used ([#1473](https://github.com/react-static/react-static/pull/1473))
+- Fix bug when calling `console.error` without arguments ([#1509](https://github.com/react-static/react-static/pull/1509))
 - Fix basepath configuration for React Router ([#1520](https://github.com/react-static/react-static/pull/1520))
 - Fix mismatched chunk names between bundle and export ([#1518](https://github.com/react-static/react-static/pull/1518))
 - Bump `git-promise` to 1.0.0, fixing a security vulnerability ([#1522](https://github.com/react-static/react-static/pull/1522))
@@ -46,6 +47,7 @@
 - Fix wrong image route in production build. ([#1425](https://github.com/react-static/react-static/pull/1425))
 
 ### Bugfix
+
 - Fix static rendering
 
 ## 7.3.0